+++ src/Select/SelectInputStyle.js
@@ -8,14 +8,17 @@ export const selectInputStyle = theme => ({
   input: {
     display: 'flex',
   },
-  inputMultiline: {
-    paddingTop: '3px !important',
-    paddingBottom: '3px !important',
+  multiValueContainer: {
+    paddingLeft: '0px !important',
+  },
+  inputMultilineDense: {
+    paddingTop: '0px !important',
+    paddingBottom: '0px !important',
   },
   inputNormal: {
     display: 'flex',
-    paddingTop: '6px !important',
-    paddingBottom: '6px !important',
+    paddingTop: `${theme.spacing.unit / 2}px !important`,
+    paddingBottom: `${theme.spacing.unit / 2}px !important`,
   },
   dummyWrapper: {
     display: 'none',
@@ -24,7 +27,7 @@ export const selectInputStyle = theme => ({
     minWidth: 0,
     display: 'flex',
     flexDirection: 'column',
-    justifyContent: 'flex-end',
+    justifyContent: 'space-around',
     flexGrow: 1,
   },
   chip: {