+++ src/globalStyle.js
@@ -1,5 +1,3 @@
-import { defaultGreen, defaultOrange, defaultRed, defaultGrey } from './colors';
-
 export const globalConst = {
   minWidth: 576,
 };
@@ -7,21 +5,5 @@ export const globalConst = {
 export const globalStyle = {
   marginTop: {
     margin: 10,
-  },
-  disabled: {
-    backgroundColor: defaultGrey,
-    color: 'white',
-  },
-  buttonColorSuccess: {
-    backgroundColor: defaultGreen,
-    color: 'white',
-  },
-  buttonColorSecondary: {
-    backgroundColor: defaultOrange,
-    color: 'white',
-  },
-  buttonColorThird: {
-    backgroundColor: defaultRed,
-    color: 'white',
   },
 };